Extract state projections into shelterwood-cells
