+++ Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 default-members = [
     "crates/shelterwood",
+    "crates/shelterwood-cells",
     "crates/shelterwood-core",
     "crates/shelterwood-mailbox",
     "crates/shelterwood-runtime",
@@ -8,6 +9,7 @@ default-members = [
 ]
 members = [
     "crates/shelterwood",
+    "crates/shelterwood-cells",
     "crates/shelterwood-core",
     "crates/shelterwood-mailbox",
     "crates/shelterwood-runtime",
@@ -30,6 +32,7 @@ unreachable_pub = "warn"
 [workspace.dependencies]
 fastrand = "2.5.0"
 serde_json = "1.0.151"
+shelterwood-cells = { path = "crates/shelterwood-cells", version = "0.1.0" }
 shelterwood-core = { path = "crates/shelterwood-core", version = "0.1.0" }
 shelterwood-mailbox = { path = "crates/shelterwood-mailbox", version = "0.1.0" }
 shelterwood-runtime = { path = "crates/shelterwood-runtime", version = "0.1.0" }
